requirements.txt
pykrx>=1.0.45
yfinance>=0.2.36
feedparser>=6.0.10
gspread>=6.0.0
google-auth>=2.28.0
requests>=2.31.0
